extract ShareProjectRoot
